Remove EffectModel IDL interface
EffectModel is not a specced part of the Web Animations interface
http://w3c.github.io/web-animations/ and has not been shipped. This patch
therefore removes EffectModel and all references to it from the IDL, in
particular as a valid type for the effect argument of Element.animate(). The
EffectModel class is maintained as a part of the Web Animations implementation.

BUG=624639

Review-Url: https://codereview.chromium.org/2234683002
Cr-Original-Commit-Position: refs/heads/master@{#411239}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8e90d746cd42c66d5565b8a08220b83378a35e24