Optimize posting a WTF::Closure on WebTaskRunner
Previously the WTF:Closure (which contains a base::Closure) was wrapped
in a WebTaskRunner::Task which was then wrapped in a base::Closure!

Thanks to moving the scheduler into blink we no longer have to live with
this absurdity :)

Also rename WebTaskRunner::taskRunner() to
WebTaskRunner::toSingleThreadTaskRunner().

BUG=

Review-Url: https://codereview.chromium.org/2232253002
Cr-Original-Commit-Position: refs/heads/master@{#411353}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1fa70eceee90f421cdeaa27da46ef386cb9a04ca