Skip hit testing containers with invalid bounding boxes
When an SVG container has no children, it's bounding box is considered
invalid. We had an assert that would fire when hit testing invalid
containers, but this assert should just be an if statement because it
is okay to hit test empty containers.

BUG=629274

Review-Url: https://codereview.chromium.org/2234173002
Cr-Original-Commit-Position: refs/heads/master@{#411536}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0bb88dbb01568d8f386365abac2bfd148bb9b2b6