Fix frameset within fragment asserts in HTMLTreeBuilder
The HTMLTreeBuilder does not expect a frameset within a fragment
in the "in body" insertion mode. It seems like this is not
specifically disallowed in the spec, event though there are
other cases where the algorithm aborts here due to parsing a
fragment.

https://html.spec.whatwg.org/multipage/syntax.html#parsing-main-inbody

BUG=636057

Review-Url: https://codereview.chromium.org/2242703002
Cr-Original-Commit-Position: refs/heads/master@{#411945}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 681c95bfedb5ee83897f8a0d4104c91049ba16f6