Fix find-in-page re-scope across frame boundaries.
There is a bug that can cause a frame to be re-scoped (all matches searched again) when issuing a find next request across a frame boundary, such as from the last match in a frame, which might need to wrap back around to the first match in the same frame. If there are not many matches in that frame, this is not really noticeable except for the ticks in the tickbar sometimes disappearing for a single frame. However, on a page with a huge number of results, this can dramatically slow down the find request and you can clearly see all of the ticks in the tickbar regenerate.

This patch adjusts the conditions under which a frame is determined to need scoping in WebLocalFrameImpl::requestFind(), in order to not re-scope across frame boundaries but continue to scope in all the cases where it IS needed.

BUG=618937,627799

Review-Url: https://codereview.chromium.org/2186113002
Cr-Original-Commit-Position: refs/heads/master@{#412025}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b6f12c557fb9b1468ba6f6da5f55cc7cefc045b0