Load 4 bytes to detect BOM.
It is guaranteed that the shared buffer contains s_smallScriptThreashold (=30*1024)
bytes, but it is not guaranteed that a chunked segment holds >=2 bytes.

Before this CL, we might hit such a short segment and failed to detect
BOM of the script.
This CL copies leading 4 bytes, which is needed and enough to detect BOM,
and avoids to lack data.

BUG=634935

Review-Url: https://codereview.chromium.org/2245003002
Cr-Original-Commit-Position: refs/heads/master@{#412206}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 58cdacd743ce7c608a0284906549366a211df0f5