Fix multiple test result printing and rebaseline compositing/video/video-reflection.html
This test seems to be working fine, except for the fact
that it prints out the test result for every frame (it seems) and
the image is slightly different now. This patch modifies the test
to only act on the first seek callback, and set the image for
rebaselining.

R=ajuma@chromium.org,pdr@chromium.org
BUG=324370,636475

Review-Url: https://codereview.chromium.org/2238043002
Cr-Original-Commit-Position: refs/heads/master@{#412313}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: af0799c558d7baf6de53392287ee3af78b90e2d4