Rebaseline svg/custom/foreign-object-skew.svg on Mac
The test result looks worse than it used to, but not bad
enough to fix. It's a super marginally situation, so not
too important.

There may still be failures on Mac Debug after this. The
test previously showed Release/Debug differences. The
expectation may need to be re-added with narrowed conditions.

R=fmalita@chromium.org
BUG=390488,636475

Review-Url: https://codereview.chromium.org/2254833002
Cr-Original-Commit-Position: refs/heads/master@{#412583}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 69de2a35ac6db47eb5994b21dbbab8eecf8f3990