[css-grid] Avoid computing twice the number of auto-repeat columns
During the track sizing layout, we were computing twice the number
of auto-repeat columns.
That was causing crashes on debug and ASAN builds.
The patch computes the value once and pass it to
LayoutGrid::placeItemsOnGrid(), which avoids the issue.

Add new cases that were crashing to test introduced in r405529.

BUG=633474
TBR=svillar@igalia.com
TEST=fast/css-grid-layout/grid-indefinite-size-auto-repeat-crash.html

Review-Url: https://codereview.chromium.org/2254843002
Cr-Original-Commit-Position: refs/heads/master@{#412638}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 49bc5ba13c492f09971882e2cb842e3514d19922