DevTools security panel: switch to overview on interstitials
When an interstitial is shown, we hide the Origins sidebar because it
isn't very useful. But we also need to switch to the Overview view when
an interstitial pops up, rather than staying on the Origin view (which
might have been selected on the page before the interstitial).

BUG=638601

Review-Url: https://codereview.chromium.org/2244233008
Cr-Original-Commit-Position: refs/heads/master@{#412912}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: de2f3d3da495915676390ef210d05672fdb6f006