Rebaseline some CSS align and justify repaint tests
The test pixel results are wrong, but that's no reason to keep
expectations. We rebaseline for the current result and use the bug
to track the fact that we get it wrong. The bug exists in this case
and has been updated with the info about these tests.

TBR=wangxianzhu@chromium.org
BUG=477798,636475

Review-Url: https://codereview.chromium.org/2254953002
Cr-Original-Commit-Position: refs/heads/master@{#412915}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: aa3ce0cd1ae8063be1ec2f6f2444bedce3534aa5