Simplify sendBeacon's allowance enforcement logic
The CL http://crrev.com/408380 removed the heavy serialization
(FormData::encodeMultiPartFormData()) from BeaconFormData::serialize().
Now, it's fine that we call m_entityBody->sizeInBytes() in
BeaconFormData::size() and finish allowance checking in
sendBeaconCommon().

This CL:
- simplifies the serialize() method by removing allowance check
  from serialize().
- simplifies NavigatorBeacon by moving logic not accessing any
  NavigatorBeacon instance variable to sendBeacon().

Note: I moved the position of mixed content check from above
serialize() to below it for simplification in http://crrev.com/411630.
The original motivation of this CL was to investigate if I can preserve
the old behavior by refactoring the Beacon class. But it turned out
that serialize() is no longer heavy logic. So, I'd like to keep the
simplified code. Also, I want to note that the change hasn't impacted
the allowance enforcement logic. payloadLength is now set even when
mixed content check fails, but in NavigatorBeacon::beaconResult, the
result is applied to m_transmittedBytes only when allowed is true.

BUG=490015
R=japhet,haraken

Review-Url: https://codereview.chromium.org/2251133003
Cr-Original-Commit-Position: refs/heads/master@{#413068}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 7651cdfe9470a553579c0f8d6193d428617a06cb