Add platform/geometry pretty printers for logging and testing
This patch switches from a test-only GeometryPrinters and debug-only
toString() to a common pretty printer format for logging and testing.

For example, to print the value of a LayoutRect:
    LayoutRect overflowRect(1, 2, 3, 4);
    LOG(INFO) << "overflow: " << overflowRect.toString();
Which prints:
    overflow: "1,2 3x4"

In tests, gtests will automatically call the PrintTo functions. In
debuggers, toString() should work as expected.

Minor changes:
1) Switched floating point precision to use "%lg" which prints up to
6 significant figures, and fewer if the value is an integer.
2) Added special-cases to catch LayoutUnit's min/max/nearly{Min,Max}.

BUG=632096

Review-Url: https://codereview.chromium.org/2191233002
Cr-Original-Commit-Position: refs/heads/master@{#413347}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: bdceea3ea3872c9188c48c0a6e71d9a52ff92feb