Leak Detector: Count ScriptPromise and ActiveDOMObject on worker threads
This CL makes the leak detector count ScriptPromise and ActiveDOMObject
not only on the main thread but also on worker threads.

There was a bug that the worker threads don't wait for the termination
(crbug.com/507224) and we couldn't count those objects on worker
threads correctly, but now the bug has been fixed.

BUG=506466, 506799, 507574

Review-Url: https://codereview.chromium.org/2258873004
Cr-Original-Commit-Position: refs/heads/master@{#413474}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f3d1a58111d4763fb364d39498e2053cada1c50d