scheduler: Reset gesture state at start of a mouse drag
This patch fixes a bug where the scheduler did not reset gesture
tracking state when a new mouse drag gesture was started. This meant
that an unrelated gesture (e.g., a wheel scroll) could be used to
compute the policy for the mouse drag, potentially causing tasks
necessary to the gesture (e.g., timers) to get deferred.

We now consider a left button press as the start of a new gesture and
also keep track of which thread is processing the mouse move events
as we do with touch gestures.

BUG=638212

Review-Url: https://codereview.chromium.org/2266883002
Cr-Original-Commit-Position: refs/heads/master@{#413475}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1329d0ac5e235bd6a94ef3ac22947ea6acf67363