Extract methods that invoke git-cl to Git class.
The main purpose of this CL is to make it easier to re-use these
functions, for example in a script that triggers Blink try jobs
and waits for results.

The secondary reason is organization; these functions are related
to interacting with git-cl, not just related to updated w3c tests.

Review-Url: https://codereview.chromium.org/2265603003
Cr-Original-Commit-Position: refs/heads/master@{#413558}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ab79d74a57c7e149a365bd55749bc98974c627ea