Restore the collapsed spaces of the text when we copy it
When we copy a wrapped text(text + inline element), the trailing
spaces of the text are collapsed during layout. However, when
we copy the text, there is no space between the text and the
inline element such as <a>, <del>, <span>, and etc so we need
to restore the trailing space of the text.

Here are examples:
<p>My favorite browser is <del>ABC</del> <ins>Chrome</ins>!</p>
<p>Counter: <span id="counter"></span></p>
<dfn>world <kbd>Blink</kbd></dfn>
<div>Copy this area <a href="http://foo/">AVeryLongWordThatWillWrap</a></div>

BUG=318925
TEST=editing/pasteboard/preserve-trailing-space.html

Review-Url: https://codereview.chromium.org/2193033004
Cr-Original-Commit-Position: refs/heads/master@{#413572}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6c91ec8582845dcaa0cc8a4b9232cd844f6bf57c