[css-grid] Refactoring of grid items stretching logic
Current implementation of applyStretchAlignmentToChildIfNeeded computes
in both axis the variables that determine whether it's possible of not
to stretch a grid item along its grid container's block axis. We are
doing this because we need to deal with orthogonal flows.

However, we will end up using only the variables associated with the
grid item's direction parallel to the grid container's block-flow
direction.

This patch uses functions instead of variables so that we can compute
only the relevant variables, determining first the grid container's
block-flow direction.

Review-Url: https://codereview.chromium.org/2261363002
Cr-Original-Commit-Position: refs/heads/master@{#413594}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6be5ad0a4f54e50032a2f5628979f1969a7e4b51