Adds commands to get and set the frame size of a DevTools target.
This is a first step towards more flexible screenshots
(bit.ly/sized-screenshots): The size of the CompositorFrame sent from the
renderer and the screenshot size are determined by the size of the RWHV.
Combined with renderer-side Emulation overrides, this will allow us to obtain
screenshots of a configurable size.

We are planning on adding an additional Emulation override command that
allows us to scale and position the page content, which will enable
screenshots of a custom area of the page. Work-in-progress patch:
https://codereview.chromium.org/2237433004/

BUG=625577

Review-Url: https://codereview.chromium.org/2226323002
Cr-Original-Commit-Position: refs/heads/master@{#413721}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 10bf0b24f8656aa2046ad92ecda2d6fc78e8f7f7